shift cache
